general: Add mime type support for archives
Until now archives were managed only if activated from Nautilus itself
and if a setting was set.

There are two main problems with this.
1- Archives opened in other apps cannot be handled by Nautilus
2- Users cannot use the regular myme type handling for setting Nautilus
as the app handling archives, or unsetting it.

This patch add support for archives myme types handled by gnome-autoar
and removes the UI and setting used in the previous version.